fix_last_bug

+++ main.h
@@ -10,8 +10,8 @@
 
 
 int vprint(const char *format, va_list args);
-void unsnumtostr(int number, int base, char *buffer);
-void number_to_string(int number, int base, char *buffer);
+void unsnumtostr(long int number, int base, char *buffer);
+void number_to_string(long int number, int base, char *buffer);
 int _printf(const char *format, ...);
 int switchi(const char *format, va_list args);
 int sstr( va_list args);
